scripts/requirements.txt
google-generativeai>=0.7.0
google-cloud-texttospeech>=2.16.0
pydub>=0.25.1
